refactor(trade): split trade.go into domain-focused files
Move run functions and their helpers from the 1124-line trade.go into
four domain-focused files: trade_list.go (list, summary, fetch),
trade_sentiment.go (sentiment calculation, ETF classification),
trade_clusters.go (clusters, cluster bombs, alerts), and
trade_levels.go (levels, level touches).

trade.go retains types, constants, constructors, preset defaults,
shared filter builders, and flag helpers (~588 lines). No function
renames, logic changes, or signature changes. Staticcheck QF1003 fix
applied to the moved tradeSentimentDayAccumulator.add method (if/else
to switch).
