+++ internal/cli/trade/trade.go
@@ -2,20 +2,12 @@ package trade
 
 import (
 	"context"
-	"fmt"
-	"log/slog"
-	"maps"
-	"sort"
 	"strconv"
-	"strings"
 
 	"github.com/leodido/structcli"
 	"github.com/spf13/cobra"
 
 	"github.com/major/volumeleaders-agent/internal/cli/common"
-	"github.com/major/volumeleaders-agent/internal/client"
-	"github.com/major/volumeleaders-agent/internal/datatables"
-	"github.com/major/volumeleaders-agent/internal/models"
 )
 
 const (
@@ -574,534 +566,6 @@ func setTradeVolumeDollarFlagDefValues(cmd *cobra.Command, minDollarsDefault flo
 func setFloatFlagDefValue(cmd *cobra.Command, name string, value float64) {
 	if flag := cmd.Flags().Lookup(name); flag != nil {
 		flag.DefValue = strconv.FormatFloat(value, 'f', -1, 64)
-	}
-}
-
-func runTradeList(cmd *cobra.Command, opts *tradeListOptions) error {
-	presetName := opts.Preset
-	watchlistName := opts.Watchlist
-	tickers := common.MultiTickerValue(cmd)
-	fields, err := common.ParseJSONFieldList[models.Trade](opts.Fields)
-	if err != nil {
-		return fmt.Errorf("parsing fields flag: %w", err)
-	}
-	format, err := common.ParseOutputFormat(opts.Format)
-	if err != nil {
-		return err
-	}
-
-	lookbackDays := 0
-	if tickers != "" {
-		lookbackDays = tradeListTickerLookbackDays
-	}
-	startDate, endDate, err := common.ResolveDateRange(cmd, lookbackDays, false)
-	if err != nil {
-		return err
-	}
-
-	filters := buildTradeFilters(tradesOptionsFromListOptions(opts, tickers, startDate, endDate))
-	if presetName != "" || watchlistName != "" {
-		if presetName != "" {
-			preset, err := findPreset(presetName)
-			if err != nil {
-				return err
-			}
-			maps.Copy(filters, preset.filters)
-		}
-		if watchlistName != "" {
-			wlFilters, err := fetchWatchlistFilters(cmd.Context(), watchlistName)
-			if err != nil {
-				return err
-			}
-			maps.Copy(filters, wlFilters)
-		}
-		applyExplicitFlags(cmd, filters)
-	}
-	if tickers != "" {
-		filters["Tickers"] = tickers
-	}
-	filters["StartDate"] = startDate
-	filters["EndDate"] = endDate
-
-	dtOpts := common.DataTableOptions{Start: opts.Start, Length: opts.Length, OrderCol: opts.OrderCol, OrderDir: opts.OrderDir, Filters: filters, Fields: fields}
-	if !opts.Summary && cmd.Flags().Changed("group-by") {
-		return fmt.Errorf("--group-by only works with --summary")
-	}
-	if opts.Summary {
-		if len(fields) > 0 {
-			return fmt.Errorf("--fields cannot be used with --summary")
-		}
-		if format != common.OutputFormatJSON {
-			return fmt.Errorf("--format cannot be used with --summary")
-		}
-		return runTradeSummary(cmd, dtOpts, opts.GroupBy, startDate, endDate)
-	}
-	if format == common.OutputFormatJSON && len(fields) == 0 {
-		return runTradeListRows(cmd, dtOpts)
-	}
-	return common.RunDataTablesCommand[models.Trade](cmd, "/Trades/GetTrades", datatables.TradeColumns, dtOpts, opts.Format, "query trades")
-}
-
-func tradesOptionsFromListOptions(opts *tradeListOptions, tickers, startDate, endDate string) *tradesOptions {
-	return &tradesOptions{tickers: tickers, startDate: startDate, endDate: endDate, minVolume: opts.MinVolume, maxVolume: opts.MaxVolume, minPrice: opts.MinPrice, maxPrice: opts.MaxPrice, minDollars: opts.MinDollars, maxDollars: opts.MaxDollars, conditions: opts.Conditions, vcd: opts.VCD, securityType: opts.SecurityType, relativeSize: opts.RelativeSize, darkPools: opts.DarkPools.Int(), sweeps: opts.Sweeps.Int(), latePrints: opts.LatePrints.Int(), sigPrints: opts.SigPrints.Int(), evenShared: opts.EvenShared.Int(), tradeRank: opts.TradeRank, rankSnapshot: opts.RankSnapshot, marketCap: opts.MarketCap, premarket: opts.Premarket.Int(), rth: opts.RTH.Int(), ah: opts.AH.Int(), opening: opts.Opening.Int(), closing: opts.Closing.Int(), phantom: opts.Phantom.Int(), offsetting: opts.Offsetting.Int(), sector: opts.Sector}
-}
-
-func tradesOptionsFromSentimentOptions(opts *tradeSentimentOptions, startDate, endDate string) *tradesOptions {
-	return &tradesOptions{startDate: startDate, endDate: endDate, minVolume: opts.MinVolume, maxVolume: opts.MaxVolume, minPrice: opts.MinPrice, maxPrice: opts.MaxPrice, minDollars: opts.MinDollars, maxDollars: opts.MaxDollars, conditions: opts.Conditions, vcd: opts.VCD, securityType: opts.SecurityType, relativeSize: opts.RelativeSize, darkPools: opts.DarkPools.Int(), sweeps: opts.Sweeps.Int(), latePrints: opts.LatePrints.Int(), sigPrints: opts.SigPrints.Int(), evenShared: opts.EvenShared.Int(), tradeRank: opts.TradeRank, rankSnapshot: opts.RankSnapshot, marketCap: opts.MarketCap, premarket: opts.Premarket.Int(), rth: opts.RTH.Int(), ah: opts.AH.Int(), opening: opts.Opening.Int(), closing: opts.Closing.Int(), phantom: opts.Phantom.Int(), offsetting: opts.Offsetting.Int()}
-}
-
-
-
-func runTradeListRows(cmd *cobra.Command, opts common.DataTableOptions) error {
-	result, err := fetchTradeList(cmd, opts)
-	if err != nil {
-		return err
-	}
-	return common.PrintJSON(cmd.OutOrStdout(), cmd.Context(), models.NewTradeListRows(result))
-}
-
-func runTradeSummary(cmd *cobra.Command, opts common.DataTableOptions, groupBy tradeSummaryGroup, startDate, endDate string) error {
-	group, err := parseTradeSummaryGroup(groupBy)
-	if err != nil {
-		return err
-	}
-	trades, err := fetchTradeList(cmd, opts)
-	if err != nil {
-		return err
-	}
-	return common.PrintJSON(cmd.OutOrStdout(), cmd.Context(), summarizeTrades(trades, group, startDate, endDate))
-}
-
-func fetchTradeList(cmd *cobra.Command, opts common.DataTableOptions) ([]models.Trade, error) {
-	ctx := cmd.Context()
-	vlClient, err := common.NewCommandClient(ctx)
-	if err != nil {
-		return nil, err
-	}
-	request := common.NewDataTablesRequest(datatables.TradeColumns, opts)
-	var result []models.Trade
-	if err := vlClient.PostDataTables(ctx, "/Trades/GetTrades", request.Encode(), &result); err != nil {
-		slog.Error("failed to query trades", "error", err)
-		return nil, fmt.Errorf("query trades: %w", err)
-	}
-	return result, nil
-}
-
-type tradeSummaryGroup string
-
-const (
-	tradeSummaryGroupTicker    tradeSummaryGroup = "ticker"
-	tradeSummaryGroupDay       tradeSummaryGroup = "day"
-	tradeSummaryGroupTickerDay tradeSummaryGroup = "ticker,day"
-)
-
-func parseTradeSummaryGroup(value tradeSummaryGroup) (tradeSummaryGroup, error) {
-	normalized := strings.ToLower(strings.ReplaceAll(strings.TrimSpace(string(value)), " ", ""))
-	switch tradeSummaryGroup(normalized) {
-	case tradeSummaryGroupTicker, tradeSummaryGroupDay, tradeSummaryGroupTickerDay:
-		return tradeSummaryGroup(normalized), nil
-	default:
-		return "", fmt.Errorf("invalid group-by %q; valid values: ticker, day, ticker,day", value)
-	}
-}
-
-type tradeGroupAccumulator struct {
-	trades                 int
-	dollars                float64
-	dollarsMultiplier      float64
-	darkPool, sweep        int
-	cumulativeDistribution float64
-}
-
-func summarizeTrades(trades []models.Trade, group tradeSummaryGroup, startDate, endDate string) models.TradeSummary {
-	summary := models.TradeSummary{DateRange: models.TradeSummaryDateRange{Start: startDate, End: endDate}}
-	groups := make(map[string]tradeGroupAccumulator)
-	keyFunc := tradeSummaryKeyFunc(group)
-	for i := range trades {
-		trade := &trades[i]
-		summary.TotalTrades++
-		summary.TotalDollars += trade.Dollars
-		addTradeSummaryGroup(groups, keyFunc(trade), trade)
-	}
-	switch group {
-	case tradeSummaryGroupTicker:
-		summary.ByTicker = summarizeTradeGroups(groups)
-	case tradeSummaryGroupDay:
-		summary.ByDay = summarizeTradeGroups(groups)
-	case tradeSummaryGroupTickerDay:
-		summary.ByTickerDay = summarizeTradeGroups(groups)
-	}
-	return summary
-}
-
-func summarizeTradeGroups(groups map[string]tradeGroupAccumulator) map[string]models.TradeGroupSummary {
-	summaries := make(map[string]models.TradeGroupSummary, len(groups))
-	for key, acc := range groups {
-		summaries[key] = acc.summary()
-	}
-	return summaries
-}
-
-func tradeSummaryKeyFunc(group tradeSummaryGroup) func(*models.Trade) string {
-	switch group {
-	case tradeSummaryGroupDay:
-		return tradeDayKey
-	case tradeSummaryGroupTickerDay:
-		return tradeTickerDayKey
-	default:
-		return tradeTickerKey
-	}
-}
-
-func addTradeSummaryGroup(groups map[string]tradeGroupAccumulator, key string, trade *models.Trade) {
-	acc := groups[key]
-	acc.trades++
-	acc.dollars += trade.Dollars
-	acc.dollarsMultiplier += trade.DollarsMultiplier
-	acc.cumulativeDistribution += trade.CumulativeDistribution
-	if bool(trade.DarkPool) {
-		acc.darkPool++
-	}
-	if bool(trade.Sweep) {
-		acc.sweep++
-	}
-	groups[key] = acc
-}
-
-func (acc tradeGroupAccumulator) summary() models.TradeGroupSummary {
-	if acc.trades == 0 {
-		return models.TradeGroupSummary{}
-	}
-	count := float64(acc.trades)
-	return models.TradeGroupSummary{
-		Trades: acc.trades, Dollars: acc.dollars,
-		AvgDollarsMultiplier:      acc.dollarsMultiplier / count,
-		PctDarkPool:               float64(acc.darkPool) / count * 100,
-		PctSweep:                  float64(acc.sweep) / count * 100,
-		AvgCumulativeDistribution: acc.cumulativeDistribution / count,
-	}
-}
-
-func tradeTickerKey(trade *models.Trade) string {
-	if trade.Ticker == "" {
-		return "unknown"
-	}
-	return trade.Ticker
-}
-
-func tradeDayKey(trade *models.Trade) string {
-	if !trade.Date.Valid {
-		return "unknown"
-	}
-	return trade.Date.Format("2006-01-02")
-}
-
-func tradeTickerDayKey(trade *models.Trade) string {
-	return tradeTickerKey(trade) + "|" + tradeDayKey(trade)
-}
-
-func runTradeSentiment(cmd *cobra.Command, opts *tradeSentimentOptions) error {
-	format, err := common.ParseOutputFormat(opts.Format)
-	if err != nil {
-		return err
-	}
-	startDate, endDate, err := common.ResolveDateRange(cmd, 0, true)
-	if err != nil {
-		return err
-	}
-	tradeOpts := tradesOptionsFromSentimentOptions(opts, startDate, endDate)
-	tradeOpts.sector = "X B"
-	filters := buildTradeFilters(tradeOpts)
-	vlClient, err := common.NewCommandClient(cmd.Context())
-	if err != nil {
-		return err
-	}
-	trades, err := fetchTradeSentimentTrades(cmd, vlClient, common.DataTableOptions{Start: 0, Length: maxTradeRequestLength, OrderCol: 1, OrderDir: "desc", Filters: filters})
-	if err != nil {
-		return err
-	}
-	sentiment := summarizeTradeSentiment(trades, startDate, endDate)
-	if format == common.OutputFormatJSON {
-		return common.PrintJSON(cmd.OutOrStdout(), cmd.Context(), sentiment)
-	}
-	return common.PrintDataTablesResult(cmd.OutOrStdout(), cmd.Context(), flattenTradeSentiment(&sentiment), nil, format)
-}
-
-func flattenTradeSentiment(summary *models.TradeSentiment) []models.TradeSentimentRow {
-	rows := make([]models.TradeSentimentRow, 0, len(summary.Daily)+1)
-	for i := range summary.Daily {
-		day := &summary.Daily[i]
-		rows = append(rows, tradeSentimentDayRow(day))
-	}
-	rows = append(rows, tradeSentimentTotalsRow(&summary.Totals))
-	return rows
-}
-
-func tradeSentimentDayRow(day *models.TradeSentimentDay) models.TradeSentimentRow {
-	return models.TradeSentimentRow{Date: day.Date, BearTrades: day.Bear.Trades, BearDollars: day.Bear.Dollars, BearTopTickers: strings.Join(day.Bear.TopTickers, ";"), BullTrades: day.Bull.Trades, BullDollars: day.Bull.Dollars, BullTopTickers: strings.Join(day.Bull.TopTickers, ";"), Ratio: day.Ratio, Signal: day.Signal}
-}
-
-func tradeSentimentTotalsRow(totals *models.TradeSentimentTotals) models.TradeSentimentRow {
-	return models.TradeSentimentRow{Date: "total", BearTrades: totals.Bear.Trades, BearDollars: totals.Bear.Dollars, BearTopTickers: strings.Join(totals.Bear.TopTickers, ";"), BullTrades: totals.Bull.Trades, BullDollars: totals.Bull.Dollars, BullTopTickers: strings.Join(totals.Bull.TopTickers, ";"), Ratio: totals.Ratio, Signal: totals.Signal}
-}
-
-func runTradeClusters(cmd *cobra.Command, opts *tradeClustersOptions) error {
-	startDate, endDate, err := common.ResolveDateRange(cmd, 0, true)
-	if err != nil {
-		return err
-	}
-	fields, err := common.OutputFields[models.TradeCluster](opts.Fields, tradeClusterDefaultFields)
-	if err != nil {
-		return fmt.Errorf("parsing fields flag: %w", err)
-	}
-	return common.RunDataTablesCommand[models.TradeCluster](cmd, "/TradeClusters/GetTradeClusters", datatables.TradeClusterColumns, common.DataTableOptions{Start: opts.Start, Length: opts.Length, OrderCol: opts.OrderCol, OrderDir: opts.OrderDir, Fields: fields, Filters: map[string]string{"Tickers": common.MultiTickerValue(cmd), "StartDate": startDate, "EndDate": endDate, "MinVolume": common.IntStr(opts.MinVolume), "MaxVolume": common.IntStr(opts.MaxVolume), "MinPrice": common.FormatFloat(opts.MinPrice), "MaxPrice": common.FormatFloat(opts.MaxPrice), "MinDollars": common.FormatFloat(opts.MinDollars), "MaxDollars": common.FormatFloat(opts.MaxDollars), "VCD": common.IntStr(opts.VCD), "SecurityTypeKey": common.IntStr(opts.SecurityType), "RelativeSize": common.IntStr(opts.RelativeSize), "TradeClusterRank": common.IntStr(opts.TradeClusterRank), "SectorIndustry": opts.Sector}}, opts.Format, "query trade clusters")
-}
-
-func runTradeClusterBombs(cmd *cobra.Command, opts *tradeClusterBombsOptions) error {
-	startDate, endDate, err := common.ResolveDateRange(cmd, 0, true)
-	if err != nil {
-		return err
-	}
-	return common.RunDataTablesCommand[models.TradeClusterBomb](cmd, "/TradeClusterBombs/GetTradeClusterBombs", datatables.TradeClusterBombColumns, common.DataTableOptions{Start: opts.Start, Length: opts.Length, OrderCol: opts.OrderCol, OrderDir: opts.OrderDir, Filters: map[string]string{"Tickers": common.MultiTickerValue(cmd), "StartDate": startDate, "EndDate": endDate, "MinVolume": common.IntStr(opts.MinVolume), "MaxVolume": common.IntStr(opts.MaxVolume), "MinDollars": common.FormatFloat(opts.MinDollars), "MaxDollars": common.FormatFloat(opts.MaxDollars), "VCD": common.IntStr(opts.VCD), "SecurityTypeKey": common.IntStr(opts.SecurityType), "RelativeSize": common.IntStr(opts.RelativeSize), "TradeClusterBombRank": common.IntStr(opts.TradeClusterBombRank), "SectorIndustry": opts.Sector}}, opts.Format, "query trade cluster bombs")
-}
-
-func runTradeAlerts(cmd *cobra.Command, opts *tradeAlertsOptions) error {
-	return common.RunDataTablesCommand[models.TradeAlert](cmd, "/TradeAlerts/GetTradeAlerts", datatables.TradeColumns, common.DataTableOptions{Start: opts.Start, Length: opts.Length, OrderCol: opts.OrderCol, OrderDir: opts.OrderDir, Filters: map[string]string{"Date": opts.Date}}, opts.Format, "query trade alerts")
-}
-
-func runTradeClusterAlerts(cmd *cobra.Command, opts *tradeClusterAlertsOptions) error {
-	return common.RunDataTablesCommand[models.TradeClusterAlert](cmd, "/TradeClusterAlerts/GetTradeClusterAlerts", datatables.TradeClusterColumns, common.DataTableOptions{Start: opts.Start, Length: opts.Length, OrderCol: opts.OrderCol, OrderDir: opts.OrderDir, Filters: map[string]string{"Date": opts.Date}}, opts.Format, "query trade cluster alerts")
-}
-
-func runTradeLevels(cmd *cobra.Command, opts *tradeLevelsOptions) error {
-	format, err := common.ParseOutputFormat(opts.Format)
-	if err != nil {
-		return err
-	}
-	startDate, endDate, err := common.ResolveDateRange(cmd, 365, false)
-	if err != nil {
-		return err
-	}
-	fields, err := common.OutputFields[models.TradeLevel](opts.Fields, nil)
-	if err != nil {
-		return fmt.Errorf("parsing fields flag: %w", err)
-	}
-	ticker, err := common.SingleTickerValue(cmd)
-	if err != nil {
-		return err
-	}
-	levelOpts := tradeLevelOptionsFromLevelsOptions(opts, ticker, startDate, endDate)
-	dataOpts := common.DataTableOptions{Start: 0, Length: levelOpts.tradeLevelCount, OrderCol: 1, OrderDir: "desc", Filters: buildTradeLevelFilters(levelOpts), Fields: fields}
-	if format == common.OutputFormatJSON && len(fields) == 0 {
-		levels, err := fetchTradeLevels(cmd, dataOpts)
-		if err != nil {
-			return err
-		}
-		return common.PrintJSON(cmd.OutOrStdout(), cmd.Context(), models.NewTradeLevelRows(levels))
-	}
-	return common.RunDataTablesSingleRequestCommand[models.TradeLevel](cmd, "/TradeLevels/GetTradeLevels", datatables.TradeLevelColumns, dataOpts, opts.Format, "query trade levels")
-}
-
-func tradeLevelOptionsFromLevelsOptions(opts *tradeLevelsOptions, ticker, startDate, endDate string) *tradeLevelOptions {
-	return &tradeLevelOptions{ticker: ticker, startDate: startDate, endDate: endDate, minVolume: opts.MinVolume, maxVolume: opts.MaxVolume, minPrice: opts.MinPrice, maxPrice: opts.MaxPrice, minDollars: opts.MinDollars, maxDollars: opts.MaxDollars, vcd: opts.VCD, relativeSize: opts.RelativeSize, tradeLevelRank: opts.TradeLevelRank, tradeLevelCount: opts.TradeLevelCount}
-}
-
-func fetchTradeLevels(cmd *cobra.Command, opts common.DataTableOptions) ([]models.TradeLevel, error) {
-	ctx := cmd.Context()
-	vlClient, err := common.NewCommandClient(ctx)
-	if err != nil {
-		return nil, err
-	}
-	request := common.NewDataTablesRequest(datatables.TradeLevelColumns, opts)
-	var result []models.TradeLevel
-	if err := vlClient.PostDataTables(ctx, "/TradeLevels/GetTradeLevels", request.Encode(), &result); err != nil {
-		slog.Error("failed to query trade levels", "error", err)
-		return nil, fmt.Errorf("query trade levels: %w", err)
-	}
-	return result, nil
-}
-
-func runTradeLevelTouches(cmd *cobra.Command, opts *tradeLevelTouchesOptions) error {
-	startDate, endDate, err := common.ResolveDateRange(cmd, 0, true)
-	if err != nil {
-		return err
-	}
-	ticker, err := common.SingleTickerValue(cmd)
-	if err != nil {
-		return err
-	}
-	return common.RunDataTablesCommand[models.TradeLevelTouch](cmd, "/TradeLevelTouches/GetTradeLevelTouches", datatables.TradeLevelTouchColumns, common.DataTableOptions{Start: opts.Start, Length: opts.Length, OrderCol: opts.OrderCol, OrderDir: opts.OrderDir, Filters: map[string]string{"Tickers": ticker, "StartDate": startDate, "EndDate": endDate, "MinVolume": common.IntStr(opts.MinVolume), "MaxVolume": common.IntStr(opts.MaxVolume), "MinPrice": common.FormatFloat(opts.MinPrice), "MaxPrice": common.FormatFloat(opts.MaxPrice), "MinDollars": common.FormatFloat(opts.MinDollars), "MaxDollars": common.FormatFloat(opts.MaxDollars), "VCD": common.IntStr(opts.VCD), "RelativeSize": common.IntStr(opts.RelativeSize), "TradeLevelRank": common.IntStr(opts.TradeLevelRank), "Levels": common.IntStr(opts.TradeLevelCount)}}, opts.Format, "query trade level touches")
-}
-
-func fetchTradeSentimentTrades(cmd *cobra.Command, vlClient *client.Client, opts common.DataTableOptions) ([]models.Trade, error) {
-	request := common.NewDataTablesRequest(datatables.TradeColumns, opts)
-	var result []models.Trade
-	if err := vlClient.PostDataTables(cmd.Context(), "/Trades/GetTrades", request.Encode(), &result); err != nil {
-		slog.Error("failed to query trade sentiment", "error", err)
-		return nil, fmt.Errorf("query trade sentiment: %w", err)
-	}
-	return result, nil
-}
-
-type tradeSentimentAccumulator struct {
-	trades        int
-	dollars       float64
-	tickerDollars map[string]float64
-}
-
-type tradeSentimentDayAccumulator struct{ bear, bull tradeSentimentAccumulator }
-
-func summarizeTradeSentiment(trades []models.Trade, startDate, endDate string) models.TradeSentiment {
-	byDay := make(map[string]*tradeSentimentDayAccumulator)
-	var totals tradeSentimentDayAccumulator
-	for i := range trades {
-		trade := &trades[i]
-		if !trade.Date.Valid {
-			continue
-		}
-		side := classifyTradeSentimentSide(trade)
-		if side == "" {
-			continue
-		}
-		day := trade.Date.Format("2006-01-02")
-		acc, ok := byDay[day]
-		if !ok {
-			acc = &tradeSentimentDayAccumulator{}
-			byDay[day] = acc
-		}
-		acc.add(side, trade)
-		totals.add(side, trade)
-	}
-	days := make([]string, 0, len(byDay))
-	for day := range byDay {
-		days = append(days, day)
-	}
-	sort.Strings(days)
-	daily := make([]models.TradeSentimentDay, 0, len(days))
-	for _, day := range days {
-		daily = append(daily, byDay[day].summary(day))
-	}
-	return models.TradeSentiment{DateRange: models.TradeSentimentDateRange{Start: startDate, End: endDate}, Daily: daily, Totals: totals.summaryTotals()}
-}
-
-func (a *tradeSentimentDayAccumulator) add(side string, trade *models.Trade) {
-	if side == "bear" {
-		a.bear.add(trade)
-	} else if side == "bull" {
-		a.bull.add(trade)
-	}
-}
-
-func (a *tradeSentimentAccumulator) add(trade *models.Trade) {
-	if a.tickerDollars == nil {
-		a.tickerDollars = make(map[string]float64)
-	}
-	a.trades++
-	a.dollars += trade.Dollars
-	a.tickerDollars[trade.Ticker] += trade.Dollars
-}
-
-func (a tradeSentimentDayAccumulator) summary(day string) models.TradeSentimentDay {
-	ratio := tradeSentimentRatio(a.bull.dollars, a.bear.dollars)
-	return models.TradeSentimentDay{Date: day, Bear: a.bear.summary(), Bull: a.bull.summary(), Ratio: ratio, Signal: tradeSentimentSignal(ratio, a.bull.dollars, a.bear.dollars)}
-}
-
-func (a tradeSentimentDayAccumulator) summaryTotals() models.TradeSentimentTotals {
-	ratio := tradeSentimentRatio(a.bull.dollars, a.bear.dollars)
-	return models.TradeSentimentTotals{Bear: a.bear.summary(), Bull: a.bull.summary(), Ratio: ratio, Signal: tradeSentimentSignal(ratio, a.bull.dollars, a.bear.dollars)}
-}
-
-func (a tradeSentimentAccumulator) summary() models.TradeSentimentSide {
-	return models.TradeSentimentSide{Trades: a.trades, Dollars: a.dollars, TopTickers: topTradeSentimentTickers(a.tickerDollars, 3)}
-}
-
-func tradeSentimentRatio(bullDollars, bearDollars float64) *float64 {
-	if bearDollars == 0 {
-		return nil
-	}
-	ratio := bullDollars / bearDollars
-	return &ratio
-}
-
-func tradeSentimentSignal(ratio *float64, bullDollars, bearDollars float64) models.TradeSentimentSignal {
-	if ratio == nil {
-		switch {
-		case bullDollars > 0:
-			return models.TradeSentimentExtremeBull
-		case bearDollars > 0:
-			return models.TradeSentimentExtremeBear
-		default:
-			return models.TradeSentimentNeutral
-		}
-	}
-	switch {
-	case *ratio < 0.2:
-		return models.TradeSentimentExtremeBear
-	case *ratio < 0.5:
-		return models.TradeSentimentModerateBear
-	case *ratio <= 2.0:
-		return models.TradeSentimentNeutral
-	case *ratio <= 5.0:
-		return models.TradeSentimentModerateBull
-	default:
-		return models.TradeSentimentExtremeBull
-	}
-}
-
-type tradeSentimentTickerTotal struct {
-	ticker  string
-	dollars float64
-}
-
-func topTradeSentimentTickers(tickerDollars map[string]float64, limit int) []string {
-	if len(tickerDollars) == 0 {
-		return []string{}
-	}
-	totals := make([]tradeSentimentTickerTotal, 0, len(tickerDollars))
-	for ticker, dollars := range tickerDollars {
-		totals = append(totals, tradeSentimentTickerTotal{ticker: ticker, dollars: dollars})
-	}
-	sort.Slice(totals, func(i, j int) bool {
-		if totals[i].dollars == totals[j].dollars {
-			return totals[i].ticker < totals[j].ticker
-		}
-		return totals[i].dollars > totals[j].dollars
-	})
-	if len(totals) < limit {
-		limit = len(totals)
-	}
-	tickers := make([]string, 0, limit)
-	for _, total := range totals[:limit] {
-		tickers = append(tickers, total.ticker)
-	}
-	return tickers
-}
-
-func classifyTradeSentimentSide(trade *models.Trade) string {
-	fields := []string{trade.Sector, trade.Name}
-	if trade.Industry != nil {
-		fields = append(fields, *trade.Industry)
-	}
-	for _, field := range fields {
-		field = strings.ToLower(field)
-		switch {
-		case strings.Contains(field, "bear"):
-			return "bear"
-		case strings.Contains(field, "bull"):
-			return "bull"
-		}
-	}
-	return leveragedETFDirection(trade.Ticker)
-}
-
-func leveragedETFDirection(ticker string) string {
-	switch strings.ToUpper(strings.TrimSpace(ticker)) {
-	case "AAPD", "AMDD", "BERZ", "BITI", "BNKD", "BZQ", "DUST", "EDZ", "ERY", "FAZ", "HIBS", "KOLD", "LABD", "MEXZ", "MYY", "NVDD", "QID", "REK", "REW", "RXD", "SARK", "SCO", "SDD", "SDOW", "SDS", "SEF", "SH", "SMDD", "SOXS", "SPDN", "SPXU", "SPXS", "SQQQ", "SRS", "SSG", "SVIX", "TSDD", "TSLQ", "TSLS", "TZA", "UVIX", "WEBS", "YANG", "YCS", "ZSL":
-		return "bear"
-	case "AAPU", "AMDL", "BITU", "BOIL", "BRZU", "CURE", "CWEB", "DFEN", "DIG", "DPST", "DRN", "EDC", "ERX", "FAS", "FNGU", "GUSH", "HIBL", "LABU", "MIDU", "NAIL", "NVDL", "QLD", "ROM", "SOXL", "SPXL", "SSO", "TECL", "TMF", "TNA", "TQQQ", "TSLL", "TURB", "UDOW", "UMDD", "UPRO", "URTY", "USD", "UWM", "WEBL", "YINN":
-		return "bull"
-	default:
-		return ""
 	}
 }
